Parameters: Add AddSolids parameters (N, start, interval) Testing: sum instead of abs(sum)

+++ src/String.cpp
@@ -2,7 +2,7 @@
 
 
 void GetParValue(std::string line, std::string& PAR, std::string& VALUE) {
-	int i = line.find('=');
+	size_t i = line.find('=');
 	std::string par_tmp(line, 0, i);
 	PAR = boost::trim_copy(par_tmp);
 	if (i > 0) {
